fix: accept fractional value_max/value_min in z2m exposes
Some zigbee devices advertise numeric exposes with sub-integer ranges
(e.g. value_max: 3599.5 on a half-second timer). The exposeCommon struct
declared these as int, so json.Unmarshal aborted partway through the
zigbee2mqtt/bridge/devices payload and the controller never identified
the offending device — or any subsequent one in the snapshot.

Switch ValueMax/ValueMin to float64 to match the z2m schema.

+++ pkg/iot/routers/zigbee2mqtt/devices.go
@@ -88,8 +88,8 @@ type exposeCommon struct {
 	Property    string   `json:"property,omitempty"`
 	Type        string   `json:"type,omitempty"`
 	Unit        string   `json:"unit,omitempty"`
-	ValueMax    int      `json:"value_max,omitempty"`
-	ValueMin    int      `json:"value_min,omitempty"`
+	ValueMax    float64  `json:"value_max,omitempty"`
+	ValueMin    float64  `json:"value_min,omitempty"`
 	Values      []string `json:"values,omitempty"`
 }
